Added some Modals and some similar to two-factor authentication was added with user's actions

+++ src/main/java/com/stason/testing/controller/commands/implementent/admin/DeleteUserCommand.java
@@ -11,28 +11,38 @@ import java.util.List;
 public class DeleteUserCommand implements Command {
     @Override
     public String execute(HttpServletRequest request) throws UnsupportedEncodingException {
-        DaoFactory factory = DaoFactory.getInstance();
-        UserDao userDao = factory.createUserDao();
-        int id = Integer.parseInt(request.getParameter("id"));
-        if(request.getServletContext().getAttribute("logoutUsersId")!=null){
-            List<Integer> logoutUsersIdList = (List<Integer>)request.getServletContext().getAttribute("logoutUsersId");
-            boolean flag = true;
-            for(Integer idLogouted : logoutUsersIdList) {
-                if (idLogouted == id) {
-                    flag = false;
-                    break;
-                }
-            }
-            if(flag) {
-                logoutUsersIdList.add(id);
-                request.getServletContext().setAttribute("logoutUsersId",logoutUsersIdList);
-            }
+        if(request.getParameter("secretPassword")==null){
+            return "redirect:/web-application/testing/admin/showUsers";
+        }else {
+            String secretPassword = request.getParameter("secretPassword");
+            if (secretPassword.equals("delete")) { // todo добавити константний клас з паролями
+                DaoFactory factory = DaoFactory.getInstance();
+                UserDao userDao = factory.createUserDao();
+                int id = Integer.parseInt(request.getParameter("id"));
+                if(request.getServletContext().getAttribute("logoutUsersId")!=null){
+                    List<Integer> logoutUsersIdList = (List<Integer>)request.getServletContext().getAttribute("logoutUsersId");
+                    boolean flag = true;
+                    for(Integer idLogouted : logoutUsersIdList) {
+                        if (idLogouted == id) {
+                            flag = false;
+                            break;
+                        }
+                    }
+                    if(flag) {
+                        logoutUsersIdList.add(id);
+                        request.getServletContext().setAttribute("logoutUsersId",logoutUsersIdList);
+                    }
 
+                }
+                userDao.deletePassedTestByUserId(id);
+                userDao.delete(id);
+                request.getSession().setAttribute("pageNumber",request.getParameter("pageNumber"));
+                request.getSession().setAttribute("paginationParameter",request.getParameter("paginationParameter"));
+                return "redirect:/web-application/testing/admin/showUsers";
+            }else{
+                request.getSession().setAttribute("error", "Секретний код не співпадає");
+                return "redirect:/web-application/testing/admin/showUsers";
+            }
         }
-        userDao.deletePassedTestByUserId(id);
-        userDao.delete(id);
-        request.getSession().setAttribute("pageNumber",request.getParameter("pageNumber"));
-        request.getSession().setAttribute("paginationParameter",request.getParameter("paginationParameter"));
-        return "redirect:/web-application/testing/admin/showUsers";
     }
 }
